bili-daily-monitor/pyproject.toml
[project]
name = "bili-daily-monitor"
version = "0.1.0"
description = "Daily Bilibili trending video monitor with AI filtering and Discord notification"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
